Fix Unsigned to signed comparisons; Add caching to channel_prob; remove -g compile flag

+++ covq/anneal.h
@@ -16,7 +16,7 @@ extern "C" {
 #include "vector.h" /* vectorset */
 #include "channel.h" /* BSC_ERROR_PROB */
 
-void anneal(vectorset *codebook, int *count, int *cw_map, int trset_size,
+void anneal(vectorset *codebook, size_t *count, uint *cw_map, size_t trset_size,
             double error_prob);
 
 #ifdef __cplusplus
